refactor: extract unix op into separate file from iour
