+++ src/driver/iour/op.rs
@@ -1,16 +1,14 @@
-use std::io::{IoSlice, IoSliceMut};
-
 use io_uring::{
     opcode,
     squeue::Entry,
     types::{Fd, FsyncFlags},
 };
-use libc::{sockaddr_storage, socklen_t};
-use socket2::SockAddr;
+use libc::sockaddr_storage;
 
+pub use crate::driver::unix_op::*;
 use crate::{
-    buf::{AsIoSlices, AsIoSlicesMut, IntoInner, IoBuf, IoBufMut, OneOrVec},
-    driver::{OpCode, RawFd},
+    buf::{AsIoSlices, AsIoSlicesMut, IoBuf, IoBufMut},
+    driver::OpCode,
     op::*,
 };
 
@@ -41,29 +39,6 @@ impl OpCode for Sync {
                 FsyncFlags::empty()
             })
             .build()
-    }
-}
-
-/// Accept a connection.
-pub struct Accept {
-    pub(crate) fd: RawFd,
-    pub(crate) buffer: sockaddr_storage,
-    pub(crate) addr_len: socklen_t,
-}
-
-impl Accept {
-    /// Create [`Accept`].
-    pub fn new(fd: RawFd) -> Self {
-        Self {
-            fd,
-            buffer: unsafe { std::mem::zeroed() },
-            addr_len: std::mem::size_of::<sockaddr_storage>() as _,
-        }
-    }
-
-    /// Get the remote address from the inner buffer.
-    pub fn into_addr(self) -> SockAddr {
-        unsafe { SockAddr::new(self.buffer, self.addr_len) }
     }
 }
 
@@ -108,36 +83,6 @@ impl<T: AsIoSlices> OpCode for SendImpl<T> {
     }
 }
 
-/// Receive data and source address.
-pub struct RecvFromImpl<T: AsIoSlicesMut> {
-    pub(crate) fd: RawFd,
-    pub(crate) buffer: T,
-    pub(crate) addr: sockaddr_storage,
-    pub(crate) slices: OneOrVec<IoSliceMut<'static>>,
-    msg: libc::msghdr,
-}
-
-impl<T: AsIoSlicesMut> RecvFromImpl<T> {
-    /// Create [`RecvFrom`] or [`RecvFromVectored`].
-    pub fn new(fd: RawFd, buffer: T::Inner) -> Self {
-        Self {
-            fd,
-            buffer: T::new(buffer),
-            addr: unsafe { std::mem::zeroed() },
-            slices: OneOrVec::One(IoSliceMut::new(&mut [])),
-            msg: unsafe { std::mem::zeroed() },
-        }
-    }
-}
-
-impl<T: AsIoSlicesMut> IntoInner for RecvFromImpl<T> {
-    type Inner = (T, sockaddr_storage, socklen_t);
-
-    fn into_inner(self) -> Self::Inner {
-        (self.buffer, self.addr, self.msg.msg_namelen)
-    }
-}
-
 impl<T: AsIoSlicesMut> OpCode for RecvFromImpl<T> {
     #[allow(clippy::no_effect)]
     fn create_entry(&mut self) -> Entry {
@@ -152,36 +97,6 @@ impl<T: AsIoSlicesMut> OpCode for RecvFromImpl<T> {
             msg_flags: 0,
         };
         opcode::RecvMsg::new(Fd(self.fd), &mut self.msg).build()
-    }
-}
-
-/// Send data to specified address.
-pub struct SendToImpl<T: AsIoSlices> {
-    pub(crate) fd: RawFd,
-    pub(crate) buffer: T,
-    pub(crate) addr: SockAddr,
-    pub(crate) slices: OneOrVec<IoSlice<'static>>,
-    msg: libc::msghdr,
-}
-
-impl<T: AsIoSlices> SendToImpl<T> {
-    /// Create [`SendTo`] or [`SendToVectored`].
-    pub fn new(fd: RawFd, buffer: T::Inner, addr: SockAddr) -> Self {
-        Self {
-            fd,
-            buffer: T::new(buffer),
-            addr,
-            slices: OneOrVec::One(IoSlice::new(&[])),
-            msg: unsafe { std::mem::zeroed() },
-        }
-    }
-}
-
-impl<T: AsIoSlices> IntoInner for SendToImpl<T> {
-    type Inner = T;
-
-    fn into_inner(self) -> Self::Inner {
-        self.buffer
     }
 }
