+++ src/driver/unix_op.rs
@@ -1,0 +1,92 @@
+use std::io::{IoSlice, IoSliceMut};
+
+use libc::{sockaddr_storage, socklen_t};
+use socket2::SockAddr;
+
+use crate::{
+    buf::{AsIoSlices, AsIoSlicesMut, IntoInner, OneOrVec},
+    driver::RawFd,
+};
+
+/// Accept a connection.
+pub struct Accept {
+    pub(crate) fd: RawFd,
+    pub(crate) buffer: sockaddr_storage,
+    pub(crate) addr_len: socklen_t,
+}
+
+impl Accept {
+    /// Create [`Accept`].
+    pub fn new(fd: RawFd) -> Self {
+        Self {
+            fd,
+            buffer: unsafe { std::mem::zeroed() },
+            addr_len: std::mem::size_of::<sockaddr_storage>() as _,
+        }
+    }
+
+    /// Get the remote address from the inner buffer.
+    pub fn into_addr(self) -> SockAddr {
+        unsafe { SockAddr::new(self.buffer, self.addr_len) }
+    }
+}
+
+/// Receive data and source address.
+pub struct RecvFromImpl<T: AsIoSlicesMut> {
+    pub(crate) fd: RawFd,
+    pub(crate) buffer: T,
+    pub(crate) addr: sockaddr_storage,
+    pub(crate) slices: OneOrVec<IoSliceMut<'static>>,
+    pub(crate) msg: libc::msghdr,
+}
+
+impl<T: AsIoSlicesMut> RecvFromImpl<T> {
+    /// Create [`RecvFrom`] or [`RecvFromVectored`].
+    pub fn new(fd: RawFd, buffer: T::Inner) -> Self {
+        Self {
+            fd,
+            buffer: T::new(buffer),
+            addr: unsafe { std::mem::zeroed() },
+            slices: OneOrVec::One(IoSliceMut::new(&mut [])),
+            msg: unsafe { std::mem::zeroed() },
+        }
+    }
+}
+
+impl<T: AsIoSlicesMut> IntoInner for RecvFromImpl<T> {
+    type Inner = (T, sockaddr_storage, socklen_t);
+
+    fn into_inner(self) -> Self::Inner {
+        (self.buffer, self.addr, self.msg.msg_namelen)
+    }
+}
+
+/// Send data to specified address.
+pub struct SendToImpl<T: AsIoSlices> {
+    pub(crate) fd: RawFd,
+    pub(crate) buffer: T,
+    pub(crate) addr: SockAddr,
+    pub(crate) slices: OneOrVec<IoSlice<'static>>,
+    pub(crate) msg: libc::msghdr,
+}
+
+impl<T: AsIoSlices> SendToImpl<T> {
+    /// Create [`SendTo`] or [`SendToVectored`].
+    pub fn new(fd: RawFd, buffer: T::Inner, addr: SockAddr) -> Self {
+        Self {
+            fd,
+            buffer: T::new(buffer),
+            addr,
+            slices: OneOrVec::One(IoSlice::new(&[])),
+            msg: unsafe { std::mem::zeroed() },
+        }
+    }
+}
+
+impl<T: AsIoSlices> IntoInner for SendToImpl<T> {
+    type Inner = T;
+
+    fn into_inner(self) -> Self::Inner {
+        self.buffer
+    }
+}
